+++ Source/AllOverIt.ReactiveUI.Wpf/Extensions/ServiceCollectionExtensions.cs
@@ -5,7 +5,7 @@ namespace AllOverIt.ReactiveUI.Wpf.Extensions
 {
     public static class ServiceCollectionExtensions
     {
-        public static IServiceCollection RegisterViewTransient<TViewModel, TView>(this IServiceCollection services)
+        public static IServiceCollection RegisterWindowTransient<TViewModel, TView>(this IServiceCollection services)
             where TViewModel : class
             where TView : ReactiveWindow<TViewModel>
         {
@@ -15,7 +15,7 @@ namespace AllOverIt.ReactiveUI.Wpf.Extensions
             return services;
         }
 
-        public static IServiceCollection RegisterViewScoped<TViewModel, TView>(this IServiceCollection services)
+        public static IServiceCollection RegisterWindowScoped<TViewModel, TView>(this IServiceCollection services)
             where TViewModel : class
             where TView : ReactiveWindow<TViewModel>
         {
@@ -25,9 +25,41 @@ namespace AllOverIt.ReactiveUI.Wpf.Extensions
             return services;
         }
 
-        public static IServiceCollection RegisterViewSingleton<TViewModel, TView>(this IServiceCollection services)
+        public static IServiceCollection RegisterWindowSingleton<TViewModel, TView>(this IServiceCollection services)
             where TViewModel : class
             where TView : ReactiveWindow<TViewModel>
+        {
+            services.AddSingleton<TViewModel>();
+            services.AddSingleton<IViewFor<TViewModel>, TView>();
+
+            return services;
+        }
+
+
+
+        public static IServiceCollection RegisterUserControlTransient<TViewModel, TView>(this IServiceCollection services)
+            where TViewModel : class
+            where TView : ReactiveUserControl<TViewModel>
+        {
+            services.AddTransient<TViewModel>();
+            services.AddTransient<IViewFor<TViewModel>, TView>();
+
+            return services;
+        }
+
+        public static IServiceCollection RegisterUserControlScoped<TViewModel, TView>(this IServiceCollection services)
+            where TViewModel : class
+            where TView : ReactiveUserControl<TViewModel>
+        {
+            services.AddScoped<TViewModel>();
+            services.AddScoped<IViewFor<TViewModel>, TView>();
+
+            return services;
+        }
+
+        public static IServiceCollection RegisterUserControlSingleton<TViewModel, TView>(this IServiceCollection services)
+            where TViewModel : class
+            where TView : ReactiveUserControl<TViewModel>
         {
             services.AddSingleton<TViewModel>();
             services.AddSingleton<IViewFor<TViewModel>, TView>();